Properly organize the pure encoders
