+++ prusti-viper/src/encoder/mir/pure/interpreter/state_high.rs
@@ -1,3 +1,9 @@
+// © 2022, ETH Zurich
+//
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
 use rustc_hash::FxHashMap;
 use std::{
     fmt::{self, Display},
