+++ prusti-viper/src/encoder/mir/pure/pure_functions/mod.rs
@@ -1,11 +1,15 @@
-//! Encoder of pure functions.
+// © 2022, ETH Zurich
+//
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-mod encoder;
+//! Encoders of pure functions.
+
 mod interface;
-mod interpreter;
-mod new_encoder;
+mod encoder_high;
+mod encoder_poly;
 
 pub(crate) use interface::{
     PureEncodingContext, PureFunctionEncoderInterface, PureFunctionEncoderState,
 };
-pub(crate) use interpreter::PureFunctionBackwardInterpreter;
